requirements.txt
asgiref==3.8.1
Django==5.2.3
django-environ==0.12.0
django-jalali-date==2.0.0
djangorestframework==3.16.0
jalali_core==1.0.0
jdatetime==5.2.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
django-ordered-model==3.7.4
pillow==11.2.1


gunicorn==23.0.0


pytest==8.4.1         
pytest-django==4.11.1 
freezegun==1.5.2      

# 🧑‍💻 
Pygments==2.19.1
iniconfig==2.1.0
pluggy==1.6.0


setuptools==78.1.1
wheel==0.45.1
packaging==25.0
